refactor: split InterviewCard from interview-interface
