+++ lib/types.ts
@@ -1,1 +1,10 @@
 export type StartingStep = "auth" | "instructions";
+export type QuestionState = "ready" | "recording";
+
+export interface QuizQuestion {
+  id: number;
+  title: string;
+  questionText: string;
+  instructions: string[];
+  timeLimit: number;
+}
